middle ware auth

+++ controllers/restaurantController.js
@@ -20,7 +20,8 @@ module.exports.controller = function(app) {
     })
     //post method of the route used to create all todolist
     .post(function(req, res) {
-        restaurantData.createRestaurant(req.body)
+        if(req.authenticated && res.role == "admin"){
+            restaurantData.createRestaurant(req.body)
         .then(function(todolist) {
             res.json(todolist);
         })
@@ -28,6 +29,10 @@ module.exports.controller = function(app) {
             console.log(err);
             res.status(500).json(err);
         });
+        }else{
+            res.status(401).json({message: 'You are not authorized to access this resource.'});
+        }
+        
     });
 
     //route for updating the todolist with id passed in the url
@@ -37,7 +42,8 @@ module.exports.controller = function(app) {
     })
     //get todolist by id
     .get(function(req, res) {
-        restaurantData.findRestaurantById(req.params.id)
+        if(req.authenticated){
+            restaurantData.findRestaurantById(req.params.id)
         .then(function(todolist) {
             res.json(todolist);
         })
@@ -45,10 +51,15 @@ module.exports.controller = function(app) {
             console.log(err);
             res.status(500).json(err);
         });
+        }else{
+            res.status(401).json({message: 'You are not authorized to access this resource.'});
+        }
+        
     })
     //update todolist by id
     .put(function(req, res) {
-        restaurantData.updateRestaurant(req.params.id, req.body)
+        if(req.authenticated && res.role == "admin"){
+            restaurantData.updateRestaurant(req.params.id, req.body)
         .then(function(obj) {
             res.json(obj);
         })
@@ -56,10 +67,16 @@ module.exports.controller = function(app) {
             console.log(err);
             res.json(err);
         });
+
+        }else{
+            res.status(401).json({message: 'You are not authorized to access this resource.'});
+        }
+        
     })
     //delete todolist by id
     .delete(function(req, res) {
-        restaurantData.deleteRestaurant(req.params.id)
+        if(req.authenticated && res.role == "admin"){
+            restaurantData.deleteRestaurant(req.params.id)
         .then(function(obj) {
             res.json(obj);
         })
@@ -67,6 +84,10 @@ module.exports.controller = function(app) {
             console.log(err);
             res.json(err);
         });
+        }else{
+            res.status(401).json({message: 'You are not authorized to access this resource.'});
+        }
+        
     });
 
 };
